+++ CHANGELOG.md
@@ -5,6 +5,9 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Changed
+ - Correctly parse the CallType for Cisco Bookings
+ - Can no longer use the Cisco to dial contexts with CallType unknown
 
 ## [16.0.0] - 2020-05-23
 ### Added